pyproject.toml
[project]
name = "sellmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "celery>=5.5.2",
    "django==5.2",
    "django-celery-beat>=2.8.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework==3.16.0",
    "dotenv==0.9.9",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "openai>=1.78.0",
    "psycopg2-binary==2.9.10",
    "pyngrok>=7.2.8",
    "python-dotenv==1.1.0",
    "python-json-logger>=3.3.0",
    "pytz>=2025.2",
    "redis>=6.0.0",
    "sqlparse==0.5.3",
    "tiktoken>=0.9.0",
    "tzdata==2025.2",
]
